Revert "Merge branch 'master' into Refactor_LedDevices"
This reverts commit 978ac7542c123d2a9eca593ae0852f1812a2dd7c, reversing
changes made to aa346ed61fdb803d8c46130a431d2468205a3e44.

+++ libsrc/hyperion/Hyperion.cpp
@@ -380,7 +380,7 @@ void Hyperion::setColor(const int priority, const std::vector<ColorRgb> &ledColo
 		_effectEngine->channelCleared(priority);
 
 	// create full led vector from single/multiple colors
-	size_t size = _ledString.leds().size();
+	unsigned int size = _ledString.leds().size();
 	std::vector<ColorRgb> newLedColors;
 	while (true)
 	{
